add sh script

+++ exchange-v2/test/contracts/tokens/TestERC1155WithRoyaltiesV1.sol
@@ -5,14 +5,24 @@ pragma abicoder v2;
 
 import "@openzeppelin/contracts-upgradeable/token/ERC1155/ERC1155Upgradeable.sol";
 import "@openzeppelin/contracts-upgradeable/proxy/Initializable.sol";
-import "@rarible/royalties/contracts/impl/RoyaltiesV1Impl.sol";
-import "@rarible/royalties/contracts/LibRoyaltiesV1.sol";
+import "@sipher/royalties/contracts/impl/RoyaltiesV1Impl.sol";
+import "@sipher/royalties/contracts/LibRoyaltiesV1.sol";
 
-contract TestERC1155WithRoyaltiesV1 is Initializable, RoyaltiesV1Impl, ERC1155Upgradeable {
+contract TestERC1155WithRoyaltiesV1 is
+    Initializable,
+    RoyaltiesV1Impl,
+    ERC1155Upgradeable
+{
     function initialize() public initializer {
         _registerInterface(LibRoyaltiesV1._INTERFACE_ID_FEES);
     }
-    function mint(address to, uint tokenId, LibPart.Part[] memory _fees, uint amount) external {
+
+    function mint(
+        address to,
+        uint256 tokenId,
+        LibPart.Part[] memory _fees,
+        uint256 amount
+    ) external {
         _mint(to, tokenId, amount, "");
         _saveRoyalties(tokenId, _fees);
     }
