+++ exchange-v2/test/contracts/tokens/TestERC721WithRoyaltiesV1.sol
@@ -5,14 +5,23 @@ pragma abicoder v2;
 
 import "@openzeppelin/contracts-upgradeable/token/ERC721/ERC721Upgradeable.sol";
 import "@openzeppelin/contracts-upgradeable/proxy/Initializable.sol";
-import "@rarible/royalties/contracts/impl/RoyaltiesV1Impl.sol";
-import "@rarible/royalties/contracts/LibRoyaltiesV1.sol";
+import "@sipher/royalties/contracts/impl/RoyaltiesV1Impl.sol";
+import "@sipher/royalties/contracts/LibRoyaltiesV1.sol";
 
-contract TestERC721WithRoyaltiesV1 is Initializable, RoyaltiesV1Impl, ERC721Upgradeable {
+contract TestERC721WithRoyaltiesV1 is
+    Initializable,
+    RoyaltiesV1Impl,
+    ERC721Upgradeable
+{
     function initialize() public initializer {
         _registerInterface(LibRoyaltiesV1._INTERFACE_ID_FEES);
     }
-    function mint(address to, uint tokenId, LibPart.Part[] memory _fees) external {
+
+    function mint(
+        address to,
+        uint256 tokenId,
+        LibPart.Part[] memory _fees
+    ) external {
         _mint(to, tokenId);
         _saveRoyalties(tokenId, _fees);
     }
