+++ royalties-registry/test/contracts/royalty-registry/RoyaltiesRegistryTest.sol
@@ -3,17 +3,22 @@
 pragma solidity >=0.6.9 <0.8.0;
 pragma abicoder v2;
 
-
-import "@rarible/royalties/contracts/IRoyaltiesProvider.sol";
-import "@rarible/royalties/contracts/LibPart.sol";
+import "@sipher/royalties/contracts/IRoyaltiesProvider.sol";
+import "@sipher/royalties/contracts/LibPart.sol";
 
 contract RoyaltiesRegistryTest {
-
     event getRoyaltiesTest(LibPart.Part[] royalties);
 
-    function _getRoyalties(address royaltiesTest, address token, uint tokenId) external {
+    function _getRoyalties(
+        address royaltiesTest,
+        address token,
+        uint256 tokenId
+    ) external {
         IRoyaltiesProvider withRoyalties = IRoyaltiesProvider(royaltiesTest);
-        LibPart.Part[] memory royalties = withRoyalties.getRoyalties(token, tokenId);
+        LibPart.Part[] memory royalties = withRoyalties.getRoyalties(
+            token,
+            tokenId
+        );
         emit getRoyaltiesTest(royalties);
     }
 }
