+++ royalties-registry/test/contracts/tokens/TestERC721WithRoyaltyV2OwnableUpgradeable.sol
@@ -5,16 +5,26 @@ pragma abicoder v2;
 
 import "@openzeppelin/contracts-upgradeable/token/ERC721/ERC721Upgradeable.sol";
 import "@openzeppelin/contracts-upgradeable/proxy/Initializable.sol";
-import "@rarible/royalties/contracts/impl/RoyaltiesV2Impl.sol";
-import "@rarible/royalties/contracts/LibRoyaltiesV2.sol";
+import "@sipher/royalties/contracts/impl/RoyaltiesV2Impl.sol";
+import "@sipher/royalties/contracts/LibRoyaltiesV2.sol";
 import "@openzeppelin/contracts-upgradeable/access/OwnableUpgradeable.sol";
 
-contract TestERC721WithRoyaltiesV2OwnableUpgradeable is Initializable, RoyaltiesV2Impl, ERC721Upgradeable, OwnableUpgradeable {
+contract TestERC721WithRoyaltiesV2OwnableUpgradeable is
+    Initializable,
+    RoyaltiesV2Impl,
+    ERC721Upgradeable,
+    OwnableUpgradeable
+{
     function initialize() public initializer {
         _registerInterface(LibRoyaltiesV2._INTERFACE_ID_ROYALTIES);
         __Ownable_init_unchained();
     }
-    function mint(address to, uint tokenId, LibPart.Part[] memory _fees) external {
+
+    function mint(
+        address to,
+        uint256 tokenId,
+        LibPart.Part[] memory _fees
+    ) external {
         _mint(to, tokenId);
         _saveRoyalties(tokenId, _fees);
     }
